Refactor service module imports
